+++ src/Controller/AppController.php
@@ -55,7 +55,7 @@ class AppController extends BaseController
 
         //Authorizes the current action, if this is not an admin request
         if (!$this->request->isAdmin()) {
-            $this->Auth->allow($this->request->getParam('action'));
+            $this->Auth->allow();
         }
 
         //Adds the current sort field in the whitelist of pagination
@@ -138,8 +138,12 @@ class AppController extends BaseController
      */
     public function isAuthorized($user = null)
     {
-        //By default, admins and managers can access all actions
-        return $this->Auth->isGroup(['admin', 'manager']);
+        //Only admin and managers can access admin actions
+        if ($this->request->isAdmin()) {
+            return $this->Auth->isGroup(['admin', 'manager']);
+        }
+
+        return true;
     }
 
     /**