+++ src/Model/Validation/AppValidator.php
@@ -129,7 +129,7 @@ class AppValidator extends Validator
 
         //Slug
         $this->add('slug', [
-            'maxLength' => [
+            'lengthBetween' => [
                 'last' => true,
                 'message' => __d('me_cms', 'Must be between {0} and {1} chars', 3, 100),
                 'rule' => ['lengthBetween', 3, 100],