+++ src/TestSuite/Traits/AuthMethodsTrait.php
@@ -69,7 +69,9 @@ trait AuthMethodsTrait
      */
     protected function setUserId($id)
     {
-        $this->Controller->Auth->setUser(['id' => $id]);
+        if (!empty($this->Controller)) {
+            $this->Controller->Auth->setUser(['id' => $id]);
+        }
 
         if (method_exists($this, 'session')) {
             $this->session(['Auth' => ['User' => ['id' => $id]]]);
@@ -84,7 +86,9 @@ trait AuthMethodsTrait
      */
     protected function setUserGroup($group)
     {
-        $this->Controller->Auth->setUser(['group' => ['name' => $group]]);
+        if (!empty($this->Controller)) {
+            $this->Controller->Auth->setUser(['group' => ['name' => $group]]);
+        }
 
         if (method_exists($this, 'session')) {
             $this->session(['Auth' => ['User' => ['group' => ['name' => $group]]]]);