+++ tests/TestCase/Controller/BannersControllerTest.php
@@ -1,0 +1,88 @@
+<?php
+/**
+ * This file is part of MeCms.
+ *
+ * MeCms is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * MeCms is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with MeCms.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ */
+namespace MeCms\Test\TestCase\Controller;
+
+use Cake\Cache\Cache;
+use Cake\ORM\TableRegistry;
+use Cake\TestSuite\IntegrationTestCase;
+
+/**
+ * BannersControllerTest class
+ */
+class BannersControllerTest extends IntegrationTestCase
+{
+    /**
+     * @var \MeCms\Model\Table\BannersTable
+     */
+    protected $Banners;
+
+    /**
+     * Fixtures
+     * @var array
+     */
+    public $fixtures = [
+        'plugin.me_cms.banners',
+    ];
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->Banners = TableRegistry::get('MeCms.Banners');
+
+        Cache::clear(false, $this->Banners->cache);
+    }
+
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        unset($this->Banners);
+    }
+
+    /**
+     * Tests for `open()` method
+     * @test
+     */
+    public function testOpen()
+    {
+        $banner = $this->Banners->findById(1)->first();
+
+        $this->get(['_name' => 'banner', $banner->id]);
+        $this->assertRedirect($banner->target);
+
+        //Checks the `click_count` has been incremented
+        $newClickCount = $this->Banners->findById(1)->first()->click_count;
+        $this->assertEquals($newClickCount, $banner->click_count + 1);
+    }
+}