+++ tests/TestCase/Model/Validation/AppValidatorTest.php
@@ -244,7 +244,7 @@ class AppValidatorTest extends TestCase
 
         $this->example['Posts']['slug'] = str_repeat('a', 101);
         $this->assertEquals([
-            'slug' => ['maxLength' => 'Must be between 3 and 100 chars'],
+            'slug' => ['lengthBetween' => 'Must be between 3 and 100 chars'],
         ], $this->Posts->newEntity($this->example['Posts'])->errors());
 
         foreach (['abc', str_repeat('a', 100)] as $value) {